fix merge
fixes idrinth-api-bench/issues#565
